requirements.txt
ffmpeg-python==0.2.0
pathlib2==2.3.7; python_version < '3.4'
